feat: refactor queries
