+++ apps/api/src/lessons/api/lessons.controller.ts
@@ -55,12 +55,14 @@ import {
   sortLessonFieldsOptions,
   SortLessonFieldsOptions,
 } from "../schemas/lessonQuery";
+import { AdminLessonsService } from "../adminLessons.service";
 
 @Controller("lessons")
 @UseGuards(RolesGuard)
 export class LessonsController {
   constructor(
     private readonly lessonsService: LessonsService,
+    private readonly adminLessonsService: AdminLessonsService,
     private readonly lessonItemsService: LessonItemsService,
   ) {}
 
@@ -94,7 +96,7 @@ export class LessonsController {
     const query = { filters, sort, page, perPage };
 
     return new PaginatedResponse(
-      await this.lessonsService.getAllLessons(query),
+      await this.adminLessonsService.getAllLessons(query),
     );
   }
 
@@ -124,7 +126,8 @@ export class LessonsController {
       }>
     >
   > {
-    const availableLessons = await this.lessonsService.getAvailableLessons();
+    const availableLessons =
+      await this.adminLessonsService.getAvailableLessons();
     return new BaseResponse(availableLessons);
   }
 
@@ -150,7 +153,7 @@ export class LessonsController {
   async getLessonById(
     @Param("id") id: string,
   ): Promise<BaseResponse<ShowLessonResponse>> {
-    return new BaseResponse(await this.lessonsService.getLessonById(id));
+    return new BaseResponse(await this.adminLessonsService.getLessonById(id));
   }
 
   @Post("create-lesson")
@@ -168,7 +171,7 @@ export class LessonsController {
     @Body() createLessonBody: CreateLessonBody,
     @CurrentUser("userId") userId: string,
   ): Promise<BaseResponse<{ message: string }>> {
-    await this.lessonsService.createLesson(createLessonBody, userId);
+    await this.adminLessonsService.createLesson(createLessonBody, userId);
     return new BaseResponse({ message: "Lesson created successfully" });
   }
 
@@ -192,7 +195,7 @@ export class LessonsController {
     @Query() id: UUIDType,
     @Body() body: UpdateLessonBody,
   ): Promise<BaseResponse<{ message: string }>> {
-    await this.lessonsService.updateLesson(id, body);
+    await this.adminLessonsService.updateLesson(id, body);
     return new BaseResponse({ message: "Text block updated successfully" });
   }
 
@@ -214,7 +217,7 @@ export class LessonsController {
   async addLessonToCourse(
     @Body() body: { courseId: string; lessonId: string; displayOrder?: number },
   ): Promise<BaseResponse<{ message: string }>> {
-    await this.lessonsService.addLessonToCourse(
+    await this.adminLessonsService.addLessonToCourse(
       body.courseId,
       body.lessonId,
       body.displayOrder,
@@ -235,7 +238,7 @@ export class LessonsController {
     @Param("courseId") courseId: string,
     @Param("lessonId") lessonId: string,
   ): Promise<BaseResponse<{ message: string }>> {
-    await this.lessonsService.removeLessonFromCourse(courseId, lessonId);
+    await this.adminLessonsService.removeLessonFromCourse(courseId, lessonId);
     return new BaseResponse({
       message: "Lesson removed from course successfully",
     });
