+++ apps/api/src/lessons/lessons.module.ts
@@ -3,12 +3,21 @@ import { LessonsService } from "./lessons.service";
 import { LessonsController } from "./api/lessons.controller";
 import { S3Service } from "src/file/s3.service";
 import { LessonItemsService } from "./lessonItems.service";
-import { LessonsRepository } from "./lessons.repository";
+import { LessonsRepository } from "./repositories/lessons.repository";
+import { AdminLessonsRepository } from "./repositories/adminLessons.repository";
+import { AdminLessonsService } from "./adminLessons.service";
 
 @Module({
   imports: [],
   controllers: [LessonsController],
-  providers: [LessonsService, S3Service, LessonItemsService, LessonsRepository],
+  providers: [
+    LessonsService,
+    AdminLessonsService,
+    LessonItemsService,
+    S3Service,
+    LessonsRepository,
+    AdminLessonsRepository,
+  ],
   exports: [],
 })
 export class LessonsModule {}
